Fixed bugs and refactored like the dickens
This code should work exactly the same way as the old code, except for
cases I thought were bugs or needlessly complex design.

- Moved the `if (!expression)` check to the top because, why wouldn't it
be? It's not very useful at the end of the function
- Reordered the combine check to put the truthy condition first
- Simplified the combine checks (no need to check `=== true`
explicitly)
- Changed the combine param's default to be false, since it's only used
as a boolean
- Changed the for-of loop to a call to Array#reduce
- Removed extraneous `await`s from returns (awaited promises get
re-wrapped anyway, in an async function)
- Moved the summing code to another function and used .then instead of
await because I feel like it's more readable and it looks nice